+++ package.json
@@ -15,6 +15,7 @@
     "@tanstack/react-query": "^5.40.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "ryan-modal": "^1.1.1",
     "styled-components": "^6.1.11",
     "styled-reset": "^4.5.2"
   },
