+++ src/App.tsx
@@ -1,12 +1,12 @@
 import { QueryClientProvider } from "@tanstack/react-query";
-import ProductListPage from "@pages/ProductListPage";
+import ProductFeedPage from "@pages/ProductFeedPage";
 import { queryClient } from "@server/queryClient";
 
 function App() {
   return (
     <>
       <QueryClientProvider client={queryClient}>
-        <ProductListPage />
+        <ProductFeedPage />
       </QueryClientProvider>
     </>
   );
