+++ src/components/ProductFeedPage/ProductFilterBar.tsx
@@ -39,5 +39,6 @@ const S = {
   SelectContainer: styled.div`
     display: flex;
     justify-content: space-between;
+    width: 100%;
   `,
 };
