--- src/pages/ProductListPage.tsx
@@ -1,12 +1,0 @@
-import ProductList from "@components/ProductListPage/ProductList";
-import { ErrorToastProvider } from "@contexts/errorToast/errorToastContext";
-
-const ProductListPage = () => {
-  return (
-    <ErrorToastProvider>
-      <ProductList />
-    </ErrorToastProvider>
-  );
-};
-
-export default ProductListPage;
